Moved general classes from Game to Framework dir. Fixed compilation
