+++ src/Framework/include/SceneStateMachine.hpp
@@ -3,7 +3,7 @@
 #include <memory>
 #include <unordered_map>
 
-#include "Scene/Scene.hpp"
+#include "Scene.hpp"
 #include "Window.hpp"
 
 class SceneStateMachine
