+++ src/Game/include/Components/C_Transform.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Components/Component.hpp"
+#include "Component.hpp"
 
 class C_Transform : public Component
 {
